Starting to migrate create_project to rest api
