+++ backend/api/src/presentation/http/routes/mod.rs
@@ -1,2 +1,3 @@
 pub mod graphql;
 pub mod users;
+pub mod projects;
