+++ backend/api/src/presentation/http/routes/projects/mod.rs
@@ -1,0 +1,2 @@
+pub mod projects_api;
+pub use projects_api::create_project;
