+++ backend/common/infrastructure/src/lib.rs
@@ -1,3 +1,4 @@
+
 #[macro_use]
 extern crate diesel;
